% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{ai_retrieve}
\alias{ai_retrieve}
\title{Retrieve Applicatino Insights data}
\usage{
ai_retrieve(query, applicationId, apiKey, timespan = "P1D")
}
\arguments{
\item{query}{The Application Insights Kusto query, as a string}

\item{applicationId}{The Application Insights Application ID}

\item{apiKey}{The API Key for the given Application Insights instance}

\item{timespan}{The Application Insights timespan (can be overridden in
the query)}
}
\value{
All data from the query in a data.frame
}
\description{
This file retrieves data from application insights as a data.frame
}
